move test_conv_shapetracker and some test_search util into unit test (#10812)
